requirements.txt
asgiref==3.5.2
certifi==2022.9.24
Django==4.1.1
django-elasticsearch-dsl==7.2.2
django-model-utils==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
PyJWT==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.2
urllib3==1.26.12
